Revise code and resolve undocumented behaviour
I've deeply revised a strange behaviour of the unit and discovered
an undocumented states, related to situation, when the unit is
initialized with the "endstop closed".
This code is an integration of my findings and should provide a safe
and effictive way to resolve these inconsistencies.
Also, some other minor changes were conducted, to improve and/or repair
the correct behaviour.